requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-filter==24.2
djangorestframework==3.14.0
drf-yasg==1.21.7
Faker==24.2.0
google-api-core==2.17.1
google-auth==2.28.2
google-geo-type==0.3.7
google-maps-routing==0.6.7
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
inflection==0.5.1
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
toposort==1.10
uritemplate==4.1.1
urllib3==2.2.1
